+++ src/services/pool/pool.service.ts
@@ -1,11 +1,9 @@
 import { getAddress } from 'ethers/lib/utils';
 
-import { isStable, isStablePhantom, isWstETH } from '@/composables/usePool';
+import { isStable } from '@/composables/usePool';
 import { FiatCurrency } from '@/constants/currency';
-import { bnSum, bnum } from '@/lib/utils';
-import { calcUSDPlusWeightedAPR } from '@/lib/utils/apr.helper';
+import { bnum } from '@/lib/utils';
 
-import { aaveService } from '../aave/aave.service';
 import { balancerSubgraphService } from '../balancer/subgraph/balancer-subgraph.service';
 import {
   ExcludedAddresses,
@@ -18,15 +16,13 @@ import {
   PoolToken
 } from '../balancer/subgraph/types';
 import { TokenPrices } from '../coingecko/api/price.service';
-import { lidoService } from '../lido/lido.service';
-import { PoolAPRs } from '../staking/staking-rewards.service';
 import { AprConcern } from './concerns/apr.concern';
 import LiquidityConcern from './concerns/liquidity.concern';
 
 export default class PoolService {
   pool: AnyPool;
   liquidityConcern: LiquidityConcern;
-  aprConcern: AprConcern;
+  apr: AprConcern;
 
   constructor(
     pool: AnyPool,
@@ -35,7 +31,7 @@ export default class PoolService {
   ) {
     this.pool = pool;
     this.liquidityConcern = new liquidityConcernClass(this);
-    this.aprConcern = new aprConcernClass(this.pool);
+    this.apr = new aprConcernClass(this.pool);
   }
 
   public calcTotalLiquidity(
@@ -48,10 +44,6 @@ export default class PoolService {
       console.error('Failed to calc pool liquidity:', error);
       return '0';
     }
-  }
-
-  public calcAPR(): PoolAPRs {
-    return {};
   }
 
   /**
@@ -133,80 +125,6 @@ export default class PoolService {
       this.pool,
       totalLiquidityString
     );
-  }
-
-  async calcThirdPartyAPR(
-    prices: TokenPrices,
-    currency: FiatCurrency,
-    protocolFeePercentage: number
-  ) {
-    let thirdPartyAPR = '0';
-    let thirdPartyAPRBreakdown = {};
-    if (isWstETH(this.pool)) {
-      thirdPartyAPR = await lidoService.calcStEthAPRFor(
-        this.pool,
-        protocolFeePercentage
-      );
-    } else if (isStablePhantom(this.pool.poolType)) {
-      const aaveAPR = await aaveService.calcWeightedSupplyAPRFor(
-        this.pool,
-        prices,
-        currency
-      );
-      let { total } = aaveAPR;
-      const { breakdown } = aaveAPR;
-
-      // TODO burn with fire once scalable linear pool support is added.
-      // If USD+ pool, replace aave APR with USD+
-      const usdPlusPools = {
-        '0xb973ca96a3f0d61045f53255e319aedb6ed4924000000000000000000000042f':
-          '0x1aAFc31091d93C3Ff003Cff5D2d8f7bA2e728425',
-        '0xf48f01dcb2cbb3ee1f6aab0e742c2d3941039d56000000000000000000000445':
-          '0x6933ec1CA55C06a894107860c92aCdFd2Dd8512f'
-      };
-      if (Object.keys(usdPlusPools).includes(this.pool.id)) {
-        const linearPoolAddress = usdPlusPools[this.pool.id];
-        const linearPool = this.pool.onchain?.linearPools?.[linearPoolAddress];
-        if (linearPool) {
-          const wrappedToken = linearPool.wrappedToken.address;
-          const weightedAPR = await calcUSDPlusWeightedAPR(
-            this.pool,
-            linearPool,
-            linearPoolAddress,
-            prices,
-            currency
-          );
-
-          breakdown[wrappedToken] = weightedAPR.toString();
-
-          total = bnSum(Object.values(breakdown)).toString();
-        }
-      }
-
-      thirdPartyAPR = total;
-      thirdPartyAPRBreakdown = breakdown;
-    }
-
-    return {
-      thirdPartyAPR,
-      thirdPartyAPRBreakdown
-    };
-  }
-
-  calcAPR(pastPool: Pool | undefined, protocolFeePercentage: number) {
-    if (!pastPool)
-      return bnum(this.pool.totalSwapFee)
-        .times(1 - protocolFeePercentage)
-        .dividedBy(this.pool.totalLiquidity)
-        .multipliedBy(365)
-        .toString();
-
-    const swapFees = bnum(this.pool.totalSwapFee).minus(pastPool.totalSwapFee);
-    return swapFees
-      .times(1 - protocolFeePercentage)
-      .dividedBy(this.pool.totalLiquidity)
-      .multipliedBy(365)
-      .toString();
   }
 
   calcFees(pastPool: Pool | undefined): string {
